LOG-1390: Refactor functional benchmark utility to run locally

+++ test/functional/outputs/forward_to_syslog_test.go
@@ -67,7 +67,7 @@ var _ = Describe("[LogForwarding][Syslog] Functional tests", func() {
 				}), logging.OutputTypeSyslog)
 			Expect(framework.Deploy()).To(BeNil())
 
-			var MaxLen uint64 = 40000
+			var MaxLen int = 40000
 			Expect(framework.WritesNApplicationLogsOfSize(1, MaxLen)).To(BeNil())
 			// Read line from Syslog output
 			outputlogs, err := framework.ReadApplicationLogsFrom(logging.OutputTypeSyslog)
